dev-requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
idna==2.10
multidict==5.1.0
python-dotenv==0.15.0
twitchio==1.1.0
typing-extensions==3.7.4.3
websockets==8.1
yarl==1.6.3
